requirements.txt
numpy~=1.23.5
matplotlib~=3.7.1
jupyter
opencv-python~=4.7.0.72
pillow~=9.5.0
scikit-learn
scikit-image
scipy
h5py
Flask-SocketIO==4.3.1
eventlet~=0.33.3
seaborn
pandas~=2.0.1
imageio
moviepy
torchsummary>=1.5.1
pytorch-lightning
tqdm~=4.65.0
tensorboard_logger
tensorboard
python-engineio==3.13.2
python-socketio==4.6.0
prefetch-generator
flask~=2.3.2
requests~=2.30.0
ultralytics